+++ tianshou/__init__.py
@@ -2,7 +2,7 @@
 # NOTE: Import order is important to avoid circular import errors!
 from tianshou import data, env, exploration, algorithm, trainer, utils
 
-__version__ = "2.0.0b1"
+__version__ = "2.0.0b2"
 
 
 def _register_log_config_callback() -> None:
@@ -18,10 +18,10 @@ _register_log_config_callback()
 
 
 __all__ = [
-    "env",
-    "data",
-    "utils",
     "algorithm",
-    "trainer",
+    "data",
+    "env",
     "exploration",
+    "trainer",
+    "utils",
 ]