+++ tianshou/env/atari/atari_network.py
@@ -58,11 +58,7 @@ class ScaledObsInputActionReprNet(ActionReprNetWithVectorOutput):
 
 
 class DQNet(ActionReprNetWithVectorOutput[Any]):
-    """Reference: Human-level control through deep reinforcement learning.
-
-    For advanced usage (how to customize the network), please refer to
-    :ref:`build_the_network`.
-    """
+    """Reference: Human-level control through deep reinforcement learning."""
 
     def __init__(
         self,
@@ -127,11 +123,7 @@ class DQNet(ActionReprNetWithVectorOutput[Any]):
 
 
 class C51Net(DQNet):
-    """Reference: A distributional perspective on reinforcement learning.
-
-    For advanced usage (how to customize the network), please refer to
-    :ref:`build_the_network`.
-    """
+    """Reference: A distributional perspective on reinforcement learning."""
 
     def __init__(
         self,
@@ -160,11 +152,7 @@ class C51Net(DQNet):
 
 
 class RainbowNet(DQNet):
-    """Reference: Rainbow: Combining Improvements in Deep Reinforcement Learning.
-
-    For advanced usage (how to customize the network), please refer to
-    :ref:`build_the_network`.
-    """
+    """Reference: Rainbow: Combining Improvements in Deep Reinforcement Learning."""
 
     def __init__(
         self,
@@ -221,11 +209,7 @@ class RainbowNet(DQNet):
 
 
 class QRDQNet(DQNet):
-    """Reference: Distributional Reinforcement Learning with Quantile Regression.
-
-    For advanced usage (how to customize the network), please refer to
-    :ref:`build_the_network`.
-    """
+    """Reference: Distributional Reinforcement Learning with Quantile Regression."""
 
     def __init__(
         self,