+++ tianshou/utils/logger/wandb.py
@@ -1,5 +1,4 @@
 import argparse
-import contextlib
 import logging
 import os
 from collections.abc import Callable
@@ -8,9 +7,6 @@ from torch.utils.tensorboard import SummaryWriter
 
 from tianshou.utils import BaseLogger, TensorboardLogger
 from tianshou.utils.logger.logger_base import VALID_LOG_VALS_TYPE, TRestoredData
-
-with contextlib.suppress(ImportError):
-    import wandb
 
 log = logging.getLogger(__name__)
 
@@ -62,6 +58,8 @@ class WandbLogger(BaseLogger):
         disable_stats: bool = False,
         log_dir: str | None = None,
     ) -> None:
+        import wandb
+
         super().__init__(train_interval, test_interval, update_interval, info_interval)
         self.last_save_step = -1
         self.save_interval = save_interval
@@ -83,13 +81,12 @@ class WandbLogger(BaseLogger):
                 # monitor_gym=monitor_gym,  # currently disabled until gymnasium version is bumped to >1.0.0 https://github.com/wandb/wandb/issues/7047
                 dir=log_dir,
                 config=config,  # type: ignore
-                settings=wandb.Settings(_disable_stats=disable_stats),
+                settings=wandb.Settings(x_disable_stats=disable_stats),
             )
             if not wandb.run
             else wandb.run
         )
-        # TODO: don't access private attribute!
-        self.wandb_run._label(repo="tianshou")  # type: ignore
+        self.wandb_run._label(repo="tianshou")
         self.tensorboard_logger: TensorboardLogger | None = None
         self.writer: SummaryWriter | None = None
 
@@ -141,12 +138,14 @@ class WandbLogger(BaseLogger):
         :param function save_checkpoint_fn: a hook defined by user, see trainer
             documentation for detail.
         """
+        import wandb
+
         if save_checkpoint_fn and epoch - self.last_save_step >= self.save_interval:
             self.last_save_step = epoch
             checkpoint_path = save_checkpoint_fn(epoch, env_step, update_step)
 
             checkpoint_artifact = wandb.Artifact(
-                "run_" + self.wandb_run.id + "_checkpoint",  # type: ignore
+                "run_" + self.wandb_run.id + "_checkpoint",
                 type="model",
                 metadata={
                     "save/epoch": epoch,
@@ -156,11 +155,11 @@ class WandbLogger(BaseLogger):
                 },
             )
             checkpoint_artifact.add_file(str(checkpoint_path))
-            self.wandb_run.log_artifact(checkpoint_artifact)  # type: ignore
+            self.wandb_run.log_artifact(checkpoint_artifact)
 
     def restore_data(self) -> tuple[int, int, int]:
-        checkpoint_artifact = self.wandb_run.use_artifact(  # type: ignore
-            f"run_{self.wandb_run.id}_checkpoint:latest",  # type: ignore
+        checkpoint_artifact = self.wandb_run.use_artifact(
+            f"run_{self.wandb_run.id}_checkpoint:latest",
         )
         assert checkpoint_artifact is not None, "W&B dataset artifact doesn't exist"